change agregando orden (completo ya se registra las salidas en el inventario y actualizo los almacenes)

+++ database/migrations/2024_12_16_224300_create_orders_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->date('dateOrder');
-            $table->integer('total');
+            $table->decimal('total', 8, 2);
             $table->enum('statu', ["pendiente","completado","cancelado"]);
             $table->foreignId('user_id');
             $table->foreignId('customer_id');
